Correct package of `Endpoint`
