+++ core/src/main/java/dev/gihwan/lith/core/LithBuilder.java
@@ -29,7 +29,7 @@ import static java.util.Objects.requireNonNull;
 import java.util.ArrayList;
 import java.util.List;
 
-import dev.gihwan.lith.core.route.Endpoint;
+import dev.gihwan.lith.core.endpoint.Endpoint;
 
 public final class LithBuilder {
