+++ core/src/main/java/dev/gihwan/lith/core/endpoint/package-info.java
@@ -22,36 +22,7 @@
  * SOFTWARE.
  */
 
-package dev.gihwan.lith.core.route;
+@NonNullByDefault
+package dev.gihwan.lith.core.endpoint;
 
-import static java.util.Objects.requireNonNull;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.google.common.base.MoreObjects;
-
-public final class ServiceEndpoint {
-
-    public static ServiceEndpoint of(String path) {
-        requireNonNull(path, "path");
-        return new ServiceEndpoint(path);
-    }
-
-    private final String path;
-
-    @JsonCreator
-    private ServiceEndpoint(@JsonProperty("path") String path) {
-        this.path = path;
-    }
-
-    public String path() {
-        return path;
-    }
-
-    @Override
-    public String toString() {
-        return MoreObjects.toStringHelper(this)
-                          .add("path", path)
-                          .toString();
-    }
-}
+import dev.gihwan.lith.core.annotation.NonNullByDefault;
